+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Change Log
 * Remove placa das buscas do checklist (PL-3546)
 * Modifica fechamento massivo de OS via Suporte (PL-3594)
 * Remove placa das buscas, abertura e fechamento de OSs (PL-3547)
+* Remove placa dos relatórios de OSs (PL-3549)
 
 <a name="v3.1.1"></a>
 ## Version [v3.1.1](https://github.com/luizfp/PrologWebService/compare/v3.1.0...v3.1.1) (release-date) [unreleased]